🔨 [!BREAKING CHANGE] 플러그인 구조 전면 수정
구현 로직 변경에 따른 플러그인 구조 전면 수정
