+++ src/services/getCommonPages.js
