--- src/ui/components/ProjectPage/index.jsx
@@ -1,5 +1,0 @@
-function ProjectPage() {
-  return <h1>ProjectPage</h1>;
-}
-
-export default ProjectPage;
