--- src/utils/isValidFigmaUrl.js
@@ -1,8 +1,0 @@
-const isValidFigmaUrl = figmaUrl => {
-  const figmaUrlPattern =
-    /^(?:https:\/\/)?(?:www\.)?figma\.com\/file\/([0-9a-zA-Z]{22,128})(?:\/?([^?]+)?(.*))?$/;
-
-  return figmaUrlPattern.test(figmaUrl);
-};
-
-export default isValidFigmaUrl;
